chore: cleanup and split up tests
